MissingNumber 2nd Try

+++ MissingNumber.java
@@ -4,12 +4,12 @@ public class MissingNumber {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        int n = scanner.nextInt();
+        long n = scanner.nextInt();
         scanner.nextLine();
         String numList = scanner.nextLine();
 
-        for (int i = 1; i <= n; i++) {
-            if (!numList.contains(Integer.toString(i))) {
+        for (long i = 1; i <= n; i++) {
+            if (!numList.contains(Long.toString(i))) {
                 System.out.print(i);
                 break;
             }
